Add files via upload
This was my Task which was part of the assignment back in 2019 I was asked to Create a class named Customer that will determine the monthly repayment amount due by a  customer for a product bought on credit. The class has five fields: customer name, contact  number, product price, number of months and the monthly repayment amount. Write get and set  methods for each field, except for the monthly repayment amount field. The set methods must  prompt the user to enter the values for the following fields: customer name, contact number,  product price and number of months. This class also needs a method to calculate the monthly repayment amount (product price divided by the number of months). Add a subclass named Finance_Period that will determine if a customer will pay interest or not. If  the number of months to pay for the product is greater than three, the customer will pay 25%  interest, else no interest applies. The maximum number of months to pay for the product is 12. Override the calculate_repayment() method by determining if the customer will pay interest or  not and calculate the monthly repayment amount. Create a class called Customer_Finance that contains the logic to test the two classes. Prompt the  user for data for the first object where no interest applies and display the results; then prompt the  user for data where interest is applicable and display the results.